refactor(server): extract scripts, data-exchange & server-info route groups (Refs #3502)
Continue the server.ts route-extraction effort (#3502) by moving the next
cohesive, low-coupling batch of inline handlers into routes/ modules.

Groups extracted (11 inline handlers, 109 -> 98 remaining):
- scripts (7): /scripts/test, /http/test, /scripts/import, /scripts/export,
  /scripts/dependencies, /scripts/dependencies/install, DELETE /scripts/:filename
  -> new scriptRoutes.ts. The 6 server.ts-local script helpers
  (getScriptsDirectory, resolveScriptPath, sanitizeMetadataValue,
  parseScriptMetadata, getLanguageFromExtension, scriptsEndpoint) moved into
  the module. getScriptsDirectory and the public scriptsEndpoint are exported;
  server.ts re-imports them for the dev-startup log and the app-level
  /api/scripts mount (which intentionally bypasses CSRF, unchanged).
- data exchange (3): /stats, /export, /import -> new dataExchangeRoutes.ts.
- server info (1): /server-info -> new serverInfoRoutes.ts.

All handler bodies moved verbatim; auth/validation middleware
(requirePermission, requireAdmin, express.raw) and path strings unchanged, so
every route resolves byte-identically under /api. Removed now-orphaned imports
(scriptDependencyEnv, normalizeTriggerPatterns, safeFetch, SsrfBlockedError,
getDependencyStatus, installDependencies) from server.ts.

Added scriptRoutes.test.ts, dataExchangeRoutes.test.ts and
serverInfoRoutes.test.ts covering validation and happy paths.

Verification: full Vitest suite green (6740 passed, 0 failed, 0 failed
suites), tsc clean, eslint clean on all changed files (0 new errors; server.ts
error count dropped 18 -> 16).
